+++ app/templates/scripts/docs/ts.js
